requirements.txt
aws-cdk.core==0.36.0
aws-cdk.aws_ec2==0.36.0
aws-cdk.aws_ecs==0.36.0
aws-cdk.aws_ecs_patterns==0.36.0

# Workaround for jsii#413
aws-cdk.aws-autoscaling-common==0.36.0
